embassy_nrf
Module gpio

General purpose input/output (GPIO) driver.

Structs

AnyPin

Type-erased GPIO pin

Flex

GPIO flexible pin.

Input

GPIO input driver.

Output

GPIO output driver.

Enums

Level

Digital input or output level.

OutputDrive

Drive strength settings for an output pin.

Port

A GPIO port with up to 32 pins.

Pull

Pull setting for an input.

Traits

Pin

Interface for a Pin that can be configured by an Input or Output driver, or converted to an AnyPin.