embassy_nrf::gpio
Trait Pin

pub trait Pin:
    Peripheral<P = Self>
    + Into<AnyPin>
    + Pin
    + Sized
    + 'static {
    // Provided methods
    fn pin(&self) -> u8 { ... }
    fn port(&self) -> Port { ... }
    fn psel_bits(&self) -> u32 { ... }
    fn degrade(self) -> AnyPin { ... }
}

Interface for a Pin that can be configured by an Input or Output driver, or converted to an AnyPin.

Provided Methods

fn pin(&self) -> u8

Number of the pin within the port (0..31)

fn port(&self) -> Port

Port of the pin

fn psel_bits(&self) -> u32

Peripheral port register value

fn degrade(self) -> AnyPin

Convert from concrete pin type PX_XX to type erased AnyPin.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.