embassy_nrf::interrupt
Module typelevel

Type-level interrupt infrastructure.

This module contains one type per interrupt. This is used for checking at compile time that the interrupts are correctly bound to HAL drivers.

As an end user, you shouldn’t need to use this module directly. Use the crate::bind_interrupts! macro to bind interrupts, and the [crate::interrupt] module to manually register interrupt handlers and manipulate interrupts directly (pending/unpending, enabling/disabling, setting the priority, etc…)

Enums

CCM_AAR

CCM_AAR typelevel interrupt.

COMP_LPCOMP

COMP_LPCOMP typelevel interrupt.

ECB

ECB typelevel interrupt.

FPU

FPU typelevel interrupt.

GPIOTE

GPIOTE typelevel interrupt.

I2S

I2S typelevel interrupt.

MWU

MWU typelevel interrupt.

NFCT

NFCT typelevel interrupt.

PDM

PDM typelevel interrupt.

POWER_CLOCK

POWER_CLOCK typelevel interrupt.

PWM0

PWM0 typelevel interrupt.

PWM1

PWM1 typelevel interrupt.

PWM2

PWM2 typelevel interrupt.

PWM3

PWM3 typelevel interrupt.

QDEC

QDEC typelevel interrupt.

RADIO

RADIO typelevel interrupt.

RNG

RNG typelevel interrupt.

RTC0

RTC0 typelevel interrupt.

RTC1

RTC1 typelevel interrupt.

RTC2

RTC2 typelevel interrupt.

SAADC

SAADC typelevel interrupt.

SPIM0_SPIS0_TWIM0_TWIS0_SPI0_TWI0

SPIM0_SPIS0_TWIM0_TWIS0_SPI0_TWI0 typelevel interrupt.

SPIM3

SPIM3 typelevel interrupt.

SPIM1_SPIS1_TWIM1_TWIS1_SPI1_TWI1

SPIM1_SPIS1_TWIM1_TWIS1_SPI1_TWI1 typelevel interrupt.

SPIM2_SPIS2_SPI2

SPIM2_SPIS2_SPI2 typelevel interrupt.

SWI0_EGU0

SWI0_EGU0 typelevel interrupt.

SWI1_EGU1

SWI1_EGU1 typelevel interrupt.

SWI2_EGU2

SWI2_EGU2 typelevel interrupt.

SWI3_EGU3

SWI3_EGU3 typelevel interrupt.

SWI4_EGU4

SWI4_EGU4 typelevel interrupt.

SWI5_EGU5

SWI5_EGU5 typelevel interrupt.

TEMP

TEMP typelevel interrupt.

TIMER0

TIMER0 typelevel interrupt.

TIMER1

TIMER1 typelevel interrupt.

TIMER2

TIMER2 typelevel interrupt.

TIMER3

TIMER3 typelevel interrupt.

TIMER4

TIMER4 typelevel interrupt.

UARTE0_UART0

UARTE0_UART0 typelevel interrupt.

UARTE1

UARTE1 typelevel interrupt.

USBD

USBD typelevel interrupt.

WDT

WDT typelevel interrupt.

Traits

Binding

Compile-time assertion that an interrupt has been bound to a handler.

Handler

Interrupt handler trait.

Interrupt

Type-level interrupt.