embassy_nrf::interrupt::typelevel
Trait Handler

pub trait Handler<I: Interrupt> {
    // Required method
    unsafe fn on_interrupt();
}

Interrupt handler trait.

Drivers that need to handle interrupts implement this trait. The user must ensure on_interrupt() is called every time the interrupt fires. Drivers must use use Binding to assert at compile time that the user has done so.

Required Methods

unsafe fn on_interrupt()

Interrupt handler function.

Must be called every time the I interrupt fires, synchronously from the interrupt handler context.

Safety

This function must ONLY be called from the interrupt handler for I.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.