embassy_nrf
Module ppi

Programmable Peripheral Interconnect (PPI/DPPI) driver.

The (Distributed) Programmable Peripheral Interconnect interface allows for an autonomous interoperability between peripherals through their events and tasks. There are fixed PPI channels and fully configurable ones. Fixed channels can only connect specific events to specific tasks. For fully configurable channels, it is possible to choose, via software, the event and the task that it will triggered by the event.

On nRF52 devices, there is also a fork task endpoint, where the user can configure one more task to be triggered by the same event, even fixed PPI channels have a configurable fork task.

The DPPI for nRF53 and nRF91 devices works in a different way. Every channel can support infinitely many tasks and events, but any single task or event can only be coupled with one channel.

Structs

AnyConfigurableChannel

The any configurable channel can represent any configurable channel at runtime. This can be used to have fewer generic parameters in some places.

AnyGroup

A type erased PPI group.

AnyStaticChannel

The any channel can represent any static channel at runtime. This can be used to have fewer generic parameters in some places.

Event

Represents an event that a peripheral can publish.

Ppi

PPI channel driver.

PpiGroup

PPI channel group driver.

Task

Represents a task that a peripheral can do.

Traits

Channel

Interface for PPI channels.

ConfigurableChannel

Interface for PPI channels that can be configured.

Group

Interface for a group of PPI channels.

StaticChannel

Interface for PPI channels that cannot be configured.