embassy_nrf::saadc
Trait Input

pub trait Input:
    Input
    + Into<AnyInput>
    + Peripheral<P = Self>
    + Sized
    + 'static {
    // Provided method
    fn degrade_saadc(self) -> AnyInput { ... }
}

An input that can be used as either or negative end of a ADC differential in the SAADC periperhal.

Provided Methods

fn degrade_saadc(self) -> AnyInput

Convert this SAADC input to a type-erased AnyInput.

This allows using several inputs in situations that might require them to be the same type, like putting them in an array.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.