embassy_nrf
Struct Peripherals

pub struct Peripherals {

    pub USBD: USBD,
    pub RTC0: RTC0,
    pub RTC1: RTC1,
    pub RTC2: RTC2,
    pub WDT: WDT,
    pub NVMC: NVMC,
    pub RNG: RNG,
    pub UARTE0: UARTE0,
    pub UARTE1: UARTE1,
    pub TWISPI0: TWISPI0,
    pub TWISPI1: TWISPI1,
    pub SPI2: SPI2,
    pub SPI3: SPI3,
    pub SAADC: SAADC,
    pub PWM0: PWM0,
    pub PWM1: PWM1,
    pub PWM2: PWM2,
    pub PWM3: PWM3,
    pub TIMER0: TIMER0,
    pub TIMER1: TIMER1,
    pub TIMER2: TIMER2,
    pub TIMER3: TIMER3,
    pub TIMER4: TIMER4,
    pub GPIOTE_CH0: GPIOTE_CH0,
    pub GPIOTE_CH1: GPIOTE_CH1,
    pub GPIOTE_CH2: GPIOTE_CH2,
    pub GPIOTE_CH3: GPIOTE_CH3,
    pub GPIOTE_CH4: GPIOTE_CH4,
    pub GPIOTE_CH5: GPIOTE_CH5,
    pub GPIOTE_CH6: GPIOTE_CH6,
    pub GPIOTE_CH7: GPIOTE_CH7,
    pub PPI_CH0: PPI_CH0,
    pub PPI_CH1: PPI_CH1,
    pub PPI_CH2: PPI_CH2,
    pub PPI_CH3: PPI_CH3,
    pub PPI_CH4: PPI_CH4,
    pub PPI_CH5: PPI_CH5,
    pub PPI_CH6: PPI_CH6,
    pub PPI_CH7: PPI_CH7,
    pub PPI_CH8: PPI_CH8,
    pub PPI_CH9: PPI_CH9,
    pub PPI_CH10: PPI_CH10,
    pub PPI_CH11: PPI_CH11,
    pub PPI_CH12: PPI_CH12,
    pub PPI_CH13: PPI_CH13,
    pub PPI_CH14: PPI_CH14,
    pub PPI_CH15: PPI_CH15,
    pub PPI_CH16: PPI_CH16,
    pub PPI_CH17: PPI_CH17,
    pub PPI_CH18: PPI_CH18,
    pub PPI_CH19: PPI_CH19,
    pub PPI_CH20: PPI_CH20,
    pub PPI_CH21: PPI_CH21,
    pub PPI_CH22: PPI_CH22,
    pub PPI_CH23: PPI_CH23,
    pub PPI_CH24: PPI_CH24,
    pub PPI_CH25: PPI_CH25,
    pub PPI_CH26: PPI_CH26,
    pub PPI_CH27: PPI_CH27,
    pub PPI_CH28: PPI_CH28,
    pub PPI_CH29: PPI_CH29,
    pub PPI_CH30: PPI_CH30,
    pub PPI_CH31: PPI_CH31,
    pub PPI_GROUP0: PPI_GROUP0,
    pub PPI_GROUP1: PPI_GROUP1,
    pub PPI_GROUP2: PPI_GROUP2,
    pub PPI_GROUP3: PPI_GROUP3,
    pub PPI_GROUP4: PPI_GROUP4,
    pub PPI_GROUP5: PPI_GROUP5,
    pub P0_00: P0_00,
    pub P0_01: P0_01,
    pub P0_02: P0_02,
    pub P0_03: P0_03,
    pub P0_04: P0_04,
    pub P0_05: P0_05,
    pub P0_06: P0_06,
    pub P0_07: P0_07,
    pub P0_08: P0_08,
    pub P0_09: P0_09,
    pub P0_10: P0_10,
    pub P0_11: P0_11,
    pub P0_12: P0_12,
    pub P0_13: P0_13,
    pub P0_14: P0_14,
    pub P0_15: P0_15,
    pub P0_16: P0_16,
    pub P0_17: P0_17,
    pub P0_19: P0_19,
    pub P0_20: P0_20,
    pub P0_21: P0_21,
    pub P0_22: P0_22,
    pub P0_23: P0_23,
    pub P0_24: P0_24,
    pub P0_25: P0_25,
    pub P0_26: P0_26,
    pub P0_27: P0_27,
    pub P0_28: P0_28,
    pub P0_29: P0_29,
    pub P0_30: P0_30,
    pub P0_31: P0_31,
    pub P1_00: P1_00,
    pub P1_01: P1_01,
    pub P1_02: P1_02,
    pub P1_03: P1_03,
    pub P1_04: P1_04,
    pub P1_05: P1_05,
    pub P1_06: P1_06,
    pub P1_07: P1_07,
    pub P1_08: P1_08,
    pub P1_09: P1_09,
    pub P1_10: P1_10,
    pub P1_11: P1_11,
    pub P1_12: P1_12,
    pub P1_13: P1_13,
    pub P1_14: P1_14,
    pub P1_15: P1_15,
    pub TEMP: TEMP,
    pub QDEC: QDEC,
    pub PDM: PDM,
    pub I2S: I2S,

}

Struct containing all the peripheral singletons.

To obtain the peripherals, you must initialize the HAL, by calling crate::init.

Fields
USBD: USBD

USBD peripheral

RTC0: RTC0

RTC0 peripheral

RTC1: RTC1

RTC1 peripheral

RTC2: RTC2

RTC2 peripheral

WDT: WDT

WDT peripheral

NVMC: NVMC

NVMC peripheral

RNG: RNG

RNG peripheral

UARTE0: UARTE0

UARTE0 peripheral

UARTE1: UARTE1

UARTE1 peripheral

TWISPI0: TWISPI0

TWISPI0 peripheral

TWISPI1: TWISPI1

TWISPI1 peripheral

SPI2: SPI2

SPI2 peripheral

SPI3: SPI3

SPI3 peripheral

SAADC: SAADC

SAADC peripheral

PWM0: PWM0

PWM0 peripheral

PWM1: PWM1

PWM1 peripheral

PWM2: PWM2

PWM2 peripheral

PWM3: PWM3

PWM3 peripheral

TIMER0: TIMER0

TIMER0 peripheral

TIMER1: TIMER1

TIMER1 peripheral

TIMER2: TIMER2

TIMER2 peripheral

TIMER3: TIMER3

TIMER3 peripheral

TIMER4: TIMER4

TIMER4 peripheral

GPIOTE_CH0: GPIOTE_CH0

GPIOTE_CH0 peripheral

GPIOTE_CH1: GPIOTE_CH1

GPIOTE_CH1 peripheral

GPIOTE_CH2: GPIOTE_CH2

GPIOTE_CH2 peripheral

GPIOTE_CH3: GPIOTE_CH3

GPIOTE_CH3 peripheral

GPIOTE_CH4: GPIOTE_CH4

GPIOTE_CH4 peripheral

GPIOTE_CH5: GPIOTE_CH5

GPIOTE_CH5 peripheral

GPIOTE_CH6: GPIOTE_CH6

GPIOTE_CH6 peripheral

GPIOTE_CH7: GPIOTE_CH7

GPIOTE_CH7 peripheral

PPI_CH0: PPI_CH0

PPI_CH0 peripheral

PPI_CH1: PPI_CH1

PPI_CH1 peripheral

PPI_CH2: PPI_CH2

PPI_CH2 peripheral

PPI_CH3: PPI_CH3

PPI_CH3 peripheral

PPI_CH4: PPI_CH4

PPI_CH4 peripheral

PPI_CH5: PPI_CH5

PPI_CH5 peripheral

PPI_CH6: PPI_CH6

PPI_CH6 peripheral

PPI_CH7: PPI_CH7

PPI_CH7 peripheral

PPI_CH8: PPI_CH8

PPI_CH8 peripheral

PPI_CH9: PPI_CH9

PPI_CH9 peripheral

PPI_CH10: PPI_CH10

PPI_CH10 peripheral

PPI_CH11: PPI_CH11

PPI_CH11 peripheral

PPI_CH12: PPI_CH12

PPI_CH12 peripheral

PPI_CH13: PPI_CH13

PPI_CH13 peripheral

PPI_CH14: PPI_CH14

PPI_CH14 peripheral

PPI_CH15: PPI_CH15

PPI_CH15 peripheral

PPI_CH16: PPI_CH16

PPI_CH16 peripheral

PPI_CH17: PPI_CH17

PPI_CH17 peripheral

PPI_CH18: PPI_CH18

PPI_CH18 peripheral

PPI_CH19: PPI_CH19

PPI_CH19 peripheral

PPI_CH20: PPI_CH20

PPI_CH20 peripheral

PPI_CH21: PPI_CH21

PPI_CH21 peripheral

PPI_CH22: PPI_CH22

PPI_CH22 peripheral

PPI_CH23: PPI_CH23

PPI_CH23 peripheral

PPI_CH24: PPI_CH24

PPI_CH24 peripheral

PPI_CH25: PPI_CH25

PPI_CH25 peripheral

PPI_CH26: PPI_CH26

PPI_CH26 peripheral

PPI_CH27: PPI_CH27

PPI_CH27 peripheral

PPI_CH28: PPI_CH28

PPI_CH28 peripheral

PPI_CH29: PPI_CH29

PPI_CH29 peripheral

PPI_CH30: PPI_CH30

PPI_CH30 peripheral

PPI_CH31: PPI_CH31

PPI_CH31 peripheral

PPI_GROUP0: PPI_GROUP0

PPI_GROUP0 peripheral

PPI_GROUP1: PPI_GROUP1

PPI_GROUP1 peripheral

PPI_GROUP2: PPI_GROUP2

PPI_GROUP2 peripheral

PPI_GROUP3: PPI_GROUP3

PPI_GROUP3 peripheral

PPI_GROUP4: PPI_GROUP4

PPI_GROUP4 peripheral

PPI_GROUP5: PPI_GROUP5

PPI_GROUP5 peripheral

P0_00: P0_00

P0_00 peripheral

P0_01: P0_01

P0_01 peripheral

P0_02: P0_02

P0_02 peripheral

P0_03: P0_03

P0_03 peripheral

P0_04: P0_04

P0_04 peripheral

P0_05: P0_05

P0_05 peripheral

P0_06: P0_06

P0_06 peripheral

P0_07: P0_07

P0_07 peripheral

P0_08: P0_08

P0_08 peripheral

P0_09: P0_09

P0_09 peripheral

P0_10: P0_10

P0_10 peripheral

P0_11: P0_11

P0_11 peripheral

P0_12: P0_12

P0_12 peripheral

P0_13: P0_13

P0_13 peripheral

P0_14: P0_14

P0_14 peripheral

P0_15: P0_15

P0_15 peripheral

P0_16: P0_16

P0_16 peripheral

P0_17: P0_17

P0_17 peripheral

P0_19: P0_19

P0_19 peripheral

P0_20: P0_20

P0_20 peripheral

P0_21: P0_21

P0_21 peripheral

P0_22: P0_22

P0_22 peripheral

P0_23: P0_23

P0_23 peripheral

P0_24: P0_24

P0_24 peripheral

P0_25: P0_25

P0_25 peripheral

P0_26: P0_26

P0_26 peripheral

P0_27: P0_27

P0_27 peripheral

P0_28: P0_28

P0_28 peripheral

P0_29: P0_29

P0_29 peripheral

P0_30: P0_30

P0_30 peripheral

P0_31: P0_31

P0_31 peripheral

P1_00: P1_00

P1_00 peripheral

P1_01: P1_01

P1_01 peripheral

P1_02: P1_02

P1_02 peripheral

P1_03: P1_03

P1_03 peripheral

P1_04: P1_04

P1_04 peripheral

P1_05: P1_05

P1_05 peripheral

P1_06: P1_06

P1_06 peripheral

P1_07: P1_07

P1_07 peripheral

P1_08: P1_08

P1_08 peripheral

P1_09: P1_09

P1_09 peripheral

P1_10: P1_10

P1_10 peripheral

P1_11: P1_11

P1_11 peripheral

P1_12: P1_12

P1_12 peripheral

P1_13: P1_13

P1_13 peripheral

P1_14: P1_14

P1_14 peripheral

P1_15: P1_15

P1_15 peripheral

TEMP: TEMP

TEMP peripheral

QDEC: QDEC

QDEC peripheral

PDM: PDM

PDM peripheral

I2S: I2S

I2S peripheral

Implementations

impl Peripherals

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.