embassy_nrf
Module uarte

Universal Asynchronous Receiver Transmitter (UART) driver.

The UART driver is provided in two flavors - this one and also crate::buffered_uarte::BufferedUarte. The Uarte here is useful for those use-cases where reading the UARTE peripheral is exclusively awaited on. If the Uarte is required to be awaited on with some other future, for example when using futures_util::future::select, then you should consider crate::buffered_uarte::BufferedUarte so that reads may continue while processing these other futures. If you do not then you may lose data between reads.

An advantage of the Uarte has over crate::buffered_uarte::BufferedUarte is that less memory may be used given that buffers are passed in directly to its read and write methods.

Structs

Config

UARTE config.

InterruptHandler

Interrupt handler.

Uarte

UARTE driver.

UarteRx

Receiver part of the UARTE driver.

UarteRxWithIdle

Receiver part of the UARTE driver, with read_until_idle support.

UarteTx

Transmitter part of the UARTE driver.

Enums

Baudrate

Baud rate

Error

UART error.

Parity

Parity

Traits

Instance

UARTE peripheral instance.